Fix #61 and add 755 permissions to rebuild.sh
This moves the -a/--authority flag to -q/--quorum and replaces it with
-a/--address, which accepts an IPv4 address. This allows the server
to bind to a custom IPv4 address, which is very useful on multi-
interface servers. If the flag is not given, it defaults to 0.0.0.0.
This behavior is also reflected in the manpage.

This commit also correctly marks rebuild.sh as executable.