Fix the segfault caused by an empty transmission
If someone connects to the server and then immediately terminates
the connection, the server will segfault and crash immediately.
This was caused by my attempt to use the Boost error message
for debugging purposes, but when no bytes are received, the error
message is not defined, hence the segfault. This also affects
OnioNS-client IPC.