Fix #65 and fix server-server loops
This fixes #65, which reports the inability for the Mirror to receive
new events by the Quorom/authority node. After further investigation,
I discovered that an async read and a call to ios.run() addressed
this, and I also took the opportunity to place the subscription
on a seperate thread.

This also fixes an infinite loop that occured between a Mirror and
the authority node, caused by incorrectly-formatted responses.
I have also simplified, renamed, and clarified the fields in the
JSON sent across the network.