Fix segfault when subscriber disappears
Mirrors subscribe to Quorum nodes, but when the Mirror disappears from the
network, Quorum nodes deallocate their Session properly. However, if the
Quorum node receives a Record, it will then attempt to tell the non-existent
Mirror. This results in a segfault because the Session has already
deallocated. The solution is to remove the Session from the subscriber list
before it deallocates.